Removed SEIRSAgeModel from models.py, copied MPC code to MPC.py
